Renamed the controller and refactored appropriatedly
